explicitly reject generic parameters
